c4018警告対応(1/2)
Win32ビルドで int と size_t の比較でC4018警告。

+++ sakura_core/view/CEditView.cpp
@@ -1844,7 +1844,7 @@ bool CEditView::GetSelectedData(
 		//>> 2002/04/18 Azumaiya
 
 		// メモリ確保に失敗したら抜ける
-		if( buffer->Capacity() < nBufSize ){
+		if( buffer->Capacity() < size_t(nBufSize) ){
 			return false;
 		}
